+++ src/parsers/mod.rs
@@ -1,3 +1,4 @@
+//! structs that parse input
 pub mod note;
 pub mod input;
 pub mod capture;
