+++ src/stores/mod.rs
@@ -1,3 +1,4 @@
+//! structs that stores data
 pub mod waveform;
 pub mod capture;
 pub mod note;
